+++ src/sequence.ts
@@ -1,9 +1,9 @@
 /**
- * Lets invoke independent functions with the same value
+ * Lets invoke independent functions with the same values
  * in order that they are declared.
  */
-export function sequence<V>(
-  ...fns: ReadonlyArray<(value: V) => unknown>
-): (value: V) => void {
-  return (value: V) => fns.forEach((fn) => fn(value));
+export function sequence<V extends ReadonlyArray<unknown>>(
+  ...fns: ReadonlyArray<(...values: V) => unknown>
+): (...values: V) => void {
+  return (...values: V) => fns.forEach((fn) => fn(...values));
 }