+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add default error type to `Task` class, `task`, `done` and `fail` functions.
 - Change type signature of `pipe` function to raise _unknown_ type if parameters chain is incompatible.
+- Change signature of `sequence` function to accept more that one argument.
 
 ## [0.16.0] - 2021-01-04